Extract track creation into dedicated component
