+++ Autobox.Core/Services/ITrackCreator.cs
@@ -1,0 +1,14 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+using System.Threading.Tasks;
+
+using Autobox.Core.Data;
+
+namespace Autobox.Core.Services
+{
+    public interface ITrackCreator
+    {
+        Task<Track> CreateTrackAsync(string url);
+    }
+}
