+++ Autobox.Desktop/App.xaml.cs
@@ -20,11 +20,13 @@ namespace Autobox.Desktop
         {
             Library = ServiceProvider.AddService<ITrackLibrary, TrackLibrary>(new TrackLibrary("Library"));
             Library.LoadAllAsync().Wait();
+            Creator = ServiceProvider.AddService<ITrackCreator, YouTubeTrackCreator>(new YouTubeTrackCreator(Library));
             Playlist = ServiceProvider.AddService<IPlaylistManager, AutoboxPlaylistManager>(new AutoboxPlaylistManager(Library));
         }
 
         // ##### Application singletons
         private readonly ITrackLibrary Library;
+        private readonly ITrackCreator Creator;
         private readonly IPlaylistManager Playlist;
     }
 }
